fix uninitialized versionSatisfied in dependency records
Records returned by the runtime dependency walker contain only 'name' and
'version', so the Dependency(SEXP) constructor left versionSatisfied (and
availableVersion) unset; if fillCranVersionInfo() then failed, an
indeterminate bool was serialized to the client, which could show a spurious
'required version not available' dialog. Use in-class member initializers so
every construction path starts from the documented defaults.

Also move the Dependency struct and buildCombinedInstallScript() into the
module header so they can be exercised by unit tests, and fix a few
comments that described the old behavior.
